README update for new project architecture.
1. Update README for new project architecture.
2. Add arm64 architecture support for all projects.
3. Slightly change on exchange library api interface.

+++ safeslingermessager/safeslingermessager/MessageReceiver.m
@@ -245,7 +245,7 @@
 	[request setHTTPMethod: @"POST"];
 	[request setHTTPBody: pktdata];
     
-    int index = [[_MsgNonces objectForKey:cnonce]integerValue];
+    NSInteger index = [[_MsgNonces objectForKey:cnonce]integerValue];
     
     NSOperationQueue *queue = [[NSOperationQueue alloc] init];
     [NSURLConnection sendAsynchronousRequest:request queue:queue completionHandler:^(NSURLResponse *response, NSData *data, NSError *error)
